App/menu UI

+++ sources/MyApplication/app/src/main/java/com/example/myapplication/Reauthenticate.java
@@ -47,7 +47,8 @@ public class Reauthenticate extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         EdgeToEdge.enable(this);
-        Objects.requireNonNull(getSupportActionBar()).hide();
+        if (getSupportActionBar() != null)
+            Objects.requireNonNull(getSupportActionBar()).hide();
         setContentView(R.layout.reauthenticate);
     }
 
@@ -68,11 +69,51 @@ public class Reauthenticate extends AppCompatActivity {
                     .addOnCompleteListener(task -> {
                         if (task.isSuccessful()) {
                             for (QueryDocumentSnapshot document : task.getResult()) {
-                                db.collection("users").document(
-                                        document.getId()).delete().addOnCompleteListener(
-                                        task2 -> Log.d("Firebase", "Document successful deleted"));
+                                db.collection("users").document(document.getId()).collection("mealDays").get()
+                                        .addOnCompleteListener(task1 -> {
+                                            if (task1.isSuccessful()){
+                                                for (QueryDocumentSnapshot documentSnapshot : task1.getResult()){
+                                                    db.collection("users").document(document.getId()).collection("mealDays").document(documentSnapshot.getId()).delete()
+                                                            .addOnCompleteListener(task2 -> {
+                                                                if (task2.isSuccessful()){
+                                                                    Log.d("Firebase", "Document successful deleted");
+                                                                    db.collection("users").document(
+                                                                            document.getId()).delete().addOnCompleteListener(
+                                                                            task3 -> {
+                                                                                if (task3.isSuccessful()){
+                                                                                    Log.d("Firebase", "Document successful deleted");
+                                                                                }
+                                                                                else {
+                                                                                    Log.d("Firebase", "Document delete failed" + task3.getException());
+                                                                                    Toast.makeText(this, "Document delete failed",
+                                                                                            Toast.LENGTH_SHORT).show();
+                                                                                }
+                                                                            });
+                                                                }
+                                                                else {
+                                                                    Log.d("Firebase", "Document delete failed" + task2.getException());
+                                                                    Toast.makeText(this, "Document delete failed",
+                                                                            Toast.LENGTH_SHORT).show();
+                                                                }
+                                                            });
+                                                }
+                                            }
+                                            else
+                                            {
+                                                Log.d("SELECT", "Failed to search for meals" + task1.getException());
+                                                Toast.makeText(this, "Failed to search for meals",
+                                                        Toast.LENGTH_SHORT).show();
+                                            }
+                                        }
+                                );
                             }
                         }
+                        else {
+                            Log.d("SELECT", "Failed to search for user" + task.getException());
+                            Toast.makeText(this, "Failed to search for user",
+                                    Toast.LENGTH_SHORT).show();
+                        }
+
                     });
             currentUser.reauthenticate(credential)
                     .addOnCompleteListener(task -> {
